+++ RyuSocks/Auth/UsernameAndPassword.cs
@@ -14,8 +14,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-using RyuSocks.Packets.Auth;
-using RyuSocks.Packets.Auth.UsernameAndPassword;
+using RyuSocks.Auth.Packets;
 using RyuSocks.Types;
 using System;
 using System.Collections.Generic;
@@ -30,30 +29,27 @@ namespace RyuSocks.Auth
     [AuthMethodImpl(0x02)]
     public class UsernameAndPassword : IProxyAuth
     {
-        public int WrapperLength => throw new NotImplementedException();
+        public int WrapperLength => 0;
 
         public string Username;
         public string Password;
-        public bool IsClient;
-        public Dictionary<string, string> Database
-        {
-            get;
-            set;
-        }
+        public bool IsClient = true;
+
+        public Dictionary<string, string> Database { get; set; }
 
         public bool Authenticate(ReadOnlySpan<byte> incomingPacket, out ReadOnlySpan<byte> outgoingPacket)
         {
-
             if (IsClient)
             {
                 if (incomingPacket == null)
                 {
-                    outgoingPacket = new UsernameAndPasswordRequest(Username, Password).Bytes;
+                    outgoingPacket = new UsernameAndPasswordRequest(Username, Password).AsSpan();
                     return false;
                 }
 
                 UsernameAndPasswordResponse incomingResponsePacket = new(incomingPacket.ToArray());
                 incomingResponsePacket.Validate();
+
                 if (incomingResponsePacket.Status == 0)
                 {
                     outgoingPacket = null;
@@ -64,7 +60,6 @@ namespace RyuSocks.Auth
             }
 
             UsernameAndPasswordRequest requestPacket = new(incomingPacket.ToArray());
-
             requestPacket.Validate();
 
             if (Database.TryGetValue(requestPacket.Username, out string password) &&
@@ -72,19 +67,23 @@ namespace RyuSocks.Auth
             {
                 UsernameAndPasswordResponse successResponsePacket = new()
                 {
-                    Version = Constants.UaPVersion,
+                    Version = AuthConsts.UsernameAndPasswordVersion,
                     Status = 0,
                 };
-                outgoingPacket = successResponsePacket.Bytes;
+
+                outgoingPacket = successResponsePacket.AsSpan();
+
                 return true;
             }
 
             UsernameAndPasswordResponse failureResponsePacket = new()
             {
-                Version = Constants.UaPVersion,
+                Version = AuthConsts.UsernameAndPasswordVersion,
                 Status = 1,
             };
-            outgoingPacket = failureResponsePacket.Bytes;
+
+            outgoingPacket = failureResponsePacket.AsSpan();
+
             throw new AuthenticationException("The provided credentials are invalid.");
         }